+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "notemule",
+  "version": "1.0.0",
+  "description": "",
+  "private": "true",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack",
+    "build-dev": "webpack --mode=development"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Vhynn/NoteMule.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Vhynn/NoteMule/issues"
+  },
+  "homepage": "https://github.com/Vhynn/NoteMule#readme",
+  "dependencies": {
+    "quill": "^1.3.7"
+  },
+  "devDependencies": {
+    "@types/quill": "^2.0.10",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1"
+  }
+}
