--- pwTool.js
@@ -1,39 +1,0 @@
-const pwPage = "<p class=\"Title--Password\">Password</p>\n<div>\n\t<input type=\"text\" class=\"TextBox--Password\" placeholder=\"Password will generate here\"/>\n</div>\n<button class=\"Button--Generate\">Generate Password</button>\n<div class=\"Button__Holder--Lower\">\n<button class=\"Button--Copy2\">Copy</button>\n\t<button class=\"Button--Back\">Back</button>\n</div>"
-const wordList = words.split(',');
-const pageContainer = document.querySelector('.Page__Container');
-
-document.querySelector('.Navigation--PW').addEventListener('click', LoadPwPage)
-
-function LoadPwPage(){
-    document.querySelector('.Page--Notes').style.display = "none";
-
-    pageContainer.innerHTML = pwPage;
-
-    document.querySelector('.Button--Generate').addEventListener('click', GeneratePassword);
-    document.querySelector('.Button--Copy2').addEventListener('click', CopyPassword);
-    document.querySelector('.Button--Back').addEventListener('click', LoadNotes);
-}
-
-function GenerateRandom(min, max){
-    return Math.floor(Math.random() * (max - min + 1) + min)
-}
-
-function GenerateSeperator(){
-    const symbols = ['!', '@', '#', '$', '%', '&', '*'];
-    const seperator = GenerateRandom(0, 999).toString() + symbols[GenerateRandom(0, 6)];
-
-    while (seperator.length<4){
-        seperator = '0' + seperator;
-    }
-
-    return seperator;
-}
-
-function GeneratePassword(){
-    const password = wordList[GenerateRandom(0, (wordList.length - 1))] + GenerateSeperator() + wordList[GenerateRandom(0, (wordList.length - 1))];
-    document.querySelector('.TextBox--Password').value = password;
-}
-
-function CopyPassword(){
-    navigator.clipboard.writeText(document.querySelector('.TextBox--Password').value);
-}
